requirements.txt
Django
psycopg2-binary
sentence-transformers
faiss-cpu
langchain
groq
requests
python-dotenv
pymupdf
django-htmx